+++ module/cards/combined-card.mjs
@@ -6,7 +6,7 @@ export class GRCard {
   static async GRAdd (config, msgId) {
     if (game.user.isGM) {
       let targetMsg = await game.messages.get(msgId)
-      if ((targetMsg.flags.brp.chatCard).length >=5 && targetMsg.flags.brp.cardType === 'GR') {
+      if ((targetMsg.flags.brp.chatCard).length >=5 && (targetMsg.flags.brp.cardType === 'GR' || targetMsg.flags.brp.cardType === 'CO')) {
         ui.notifications.warn(game.i18n.localize('BRP.resolveMax'))    
         return
       } else if ((targetMsg.flags.brp.chatCard).length >=2 && targetMsg.flags.brp.cardType === 'OP') {
@@ -60,8 +60,7 @@ export class GRCard {
     await roll.roll({ async: true})
     let rollResult = Number(roll.result)
 
-
-
+    let successes = 0
     for (let i of chatCards) {
       i.rollResult = rollResult
       i.rollVal = rollResult
@@ -70,22 +69,27 @@ export class GRCard {
         rollVal: i.rollVal,
         cardType,
       })
+      if (i.resultLevel > 1) {successes++}
       i.resultLabel = game.i18n.localize('BRP.resultLevel.'+i.resultLevel)
       newchatCards.push(i)
-    }    
+    }
+    successes = successes/chatCards.length    
     AudioHelper.play({ src: CONFIG.sounds.dice }, true)
     await targetMsg.update({'flags.brp.chatCard' :newchatCards,
                             'flags.brp.state': 'closed',
+                            'flags.brp.successLevel': successes,
                             'flags.brp.rollResult': rollResult,
                             'rolls': [roll]})
     const pushhtml = await BRPCheck.startChat(targetMsg.flags.brp)
     await targetMsg.update({content: pushhtml})                        
+    await BRPCheck.tickXP (targetMsg.flags.brp)
     return
   }  
 
   static async GRClose (config) {
     let targetMsg = await game.messages.get(config.targetChatId)
     await targetMsg.update({'flags.brp.state': 'closed',
+                            'flgs.brp.successLevel': -1,
                             'flags.brp.chatCard' :[]})
     const pushhtml = await BRPCheck.startChat(targetMsg.flags.brp)
     await targetMsg.update({content: pushhtml})