+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## 11.1.9
+- Added GM Tools to the scene tools and a Development Phase
+- Added XP improvement checks, only available when Development Phase is turned on by the GM.  Applies to Normal Skill, Combined, Oppossed and POW v POW rolls.
+- Added POW improvement roll, only available when Dev Phase is turned on
+- Option to turn on "auto XP" checks when succesfully using a skill
+- Cooperative skill rolls added
+- Can roll weapon damage from items in the Combat tab (unless damage is "Special")
+- Can roll weapon attacks from skill% in items in the Combat tab
+- Updated partial success rolls for Combined rolls
+
 ## 11.1.8
 - Added Socket functionality (pre-cursor to combined/opposed rolls)
 - Combined and Opposed skill rolls added